requirements.txt
beautifulsoup4==4.8.1
bs4==0.0.1
html2text==2019.9.26
mysql==0.0.2
mysql-connector==2.2.9
mysqlclient==1.4.6
PyYAML==5.1.2
requests==2.22.0
